Simplify even more
