+++ bundle/Resources/config/services.yaml
@@ -7,8 +7,5 @@ services:
     Novactive\Bundle\FormBuilderBundle\Core\:
         resource: '../../Core'
 
-    Novactive\Bundle\FormBuilderBundle\Field\Type\:
-        resource: '../../Field/Type'
-
     Novactive\Bundle\FormBuilderBundle\Core\FieldTypeRegistry:
         arguments: [!tagged novaformbuilder.fieldtype]
